+++ src/cortex/api/mental_command.py
